attempting to make a universal test
